rename scrape config entry
