+++ pkg/promtail/targets/lokipush/pushtarget_test.go
@@ -1,4 +1,4 @@
-package push
+package lokipush
 
 import (
 	"flag"
