+++ pkg/promtail/targets/lokipush/pushtargetmanager.go
@@ -1,4 +1,4 @@
-package push
+package lokipush
 
 import (
 	"errors"
