+++ pkg/promtail/targets/lokipush/pushtargetmanager_test.go
@@ -1,4 +1,4 @@
-package push
+package lokipush
 
 import (
 	"testing"
